+++ fractal_server/app/routes/admin/v2/task.py
@@ -8,6 +8,7 @@ from sqlmodel import select
 
 from fractal_server.app.db import AsyncSession
 from fractal_server.app.db import get_async_db
+from fractal_server.app.models import LinkUserProjectV2
 from fractal_server.app.models import TaskGroupV2
 from fractal_server.app.models import UserOAuth
 from fractal_server.app.models.v2 import TaskV2
@@ -133,12 +134,29 @@ async def query_tasks(
     for task in task_list:
         stm = (
             select(WorkflowV2)
-            .join(WorkflowTaskV2)
-            .where(WorkflowTaskV2.workflow_id == WorkflowV2.id)
+            .join(
+                WorkflowTaskV2,
+                WorkflowTaskV2.workflow_id == WorkflowV2.id,
+            )
             .where(WorkflowTaskV2.task_id == task.id)
         )
         res = await db.execute(stm)
         wf_list = res.scalars().all()
+
+        project_users = {}
+        for project_id in set([workflow.project_id for workflow in wf_list]):
+            res = await db.execute(
+                select(UserOAuth.id, UserOAuth.email)
+                .join(
+                    LinkUserProjectV2,
+                    LinkUserProjectV2.user_id == UserOAuth.id,
+                )
+                .where(LinkUserProjectV2.project_id == project_id)
+            )
+            project_users[project_id] = [
+                ProjectUser(id=p_user[0], email=p_user[1])
+                for p_user in res.all()
+            ]
 
         task_info_list.append(
             dict(
@@ -149,16 +167,12 @@ async def query_tasks(
                         workflow_name=workflow.name,
                         project_id=workflow.project.id,
                         project_name=workflow.project.name,
-                        project_users=[
-                            dict(id=user.id, email=user.email)
-                            for user in workflow.project.user_list
-                        ],
+                        project_users=project_users[workflow.project_id],
                     )
                     for workflow in wf_list
                 ],
             )
         )
-
     return PaginationResponse[TaskV2Info](
         total_count=total_count,
         page_size=page_size,