+++ fractal_server/app/routes/api/v2/project.py
@@ -67,12 +67,14 @@ async def create_project(
     )
 
     db_project = ProjectV2(**project.model_dump(), resource_id=resource_id)
-    db_project.user_list.append(user)
-
     db.add(db_project)
+    await db.flush()
+
+    link = LinkUserProjectV2(project_id=db_project.id, user_id=user.id)
+    db.add(link)
+
     await db.commit()
     await db.refresh(db_project)
-    await db.close()
 
     return db_project