+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     * Remove `GET /admin/v2/project/`, `GET /api/v2/dataset/` and `GET /api/v2/workflow/` (\#2989).
 * Database:
     * Add indexes to `HistoryImageCache` and `HistoryUnit` foreign keys to improve `DatasetV2` deletion time (\#2987).
+    * Drop `ProjectV2.user_list` relationship (\#2991).
 * Runner:
     * Handle default `batch_size` value for local runner (\#2949).
     * Rename `SudoSlurmRunner` into `SlurmSudoRunner` (\#2980).